+++ src/main/java/insane96mcp/iguanatweaksreborn/module/mining/MaterialsAndOres.java
@@ -7,13 +7,16 @@ import insane96mcp.insanelib.base.Feature;
 import insane96mcp.insanelib.base.LoadFeature;
 import insane96mcp.insanelib.base.Module;
 import insane96mcp.insanelib.base.config.Config;
+import net.minecraftforge.fml.ModList;
 
 @LoadFeature(module = Modules.Ids.MINING, description = "Various changes for different materials and ores.")
 public class MaterialsAndOres extends Feature {
-	@Config(description = "Enables a data pack that changes ore generation to be biome based. Less material generate in all the biomes, more in specific biomes.")
-	public static Boolean oreGenerationOverhaul = true;
 	@Config(description = "https://minecraft.wiki/w/Ore_vein")
 	public static Boolean disableOreVeins = true;
+	@Config(description = "Enables a data pack that changes ore generation to be biome based. Less material generate in all the biomes, more in specific biomes.")
+	public static Boolean oreGenerationOverhaul = true;
+	@Config(description = "Enables a data pack that backports diamonds generation from 1.20.2. Disables itself if Terralith is installed")
+	public static Boolean diamondGenerationBackport = true;
 	@Config(description = """
 			Enables the following changes to vanilla data pack:
 			* Stone (Broken with a non Silk-Touch tool) can drop Iron Nuggets
@@ -27,10 +30,11 @@ public class MaterialsAndOres extends Feature {
 			* Can no longer smelt gold and Ancient Debris in a Furnace, and 2x in a blast furnace""")
 	public static Boolean oreSmelting = true;
 
-	public MaterialsAndOres(Module module, boolean enabledByDefault, boolean canBeDisabled) {
-		super(module, enabledByDefault, canBeDisabled);
+	public void init(Module module, boolean enabledByDefault, boolean canBeDisabled) {
+		super.init(module, enabledByDefault, canBeDisabled);
 		InsaneSO.addServerPack("farmable_iron", "Insane's Survival Overhaul Farmable Iron", () -> this.isEnabled() && !DataPacks.disableAllDataPacks && farmableIronDataPack);
 		InsaneSO.addServerPack("ore_smelting", "Insane's Survival Overhaul Ore Smelting", () -> this.isEnabled() && !DataPacks.disableAllDataPacks && oreSmelting);
 		InsaneSO.addServerPack("ore_generation", "Insane's Survival Overhaul Ore Generation Overhaul", () -> this.isEnabled() && !DataPacks.disableAllDataPacks && oreGenerationOverhaul);
+		InsaneSO.addServerPack("diamond_gen_backport", "Insane's Survival Overhaul Diamond Generation Backport", () -> this.isEnabled() && !DataPacks.disableAllDataPacks && diamondGenerationBackport && !ModList.get().isLoaded("terralith"));
 	}
 }